refactor: delete tonk-concept, fold its template logic into tonk-display
With the <tonk-concept> element gone, the crate was just shared
notation-template machinery that only tonk-display used. Move template.rs
and the source-parsing/query helpers from resolve.rs into tonk-display,
retargeting the duplicate ErrorDetail/ErrorKind at tonk_host::error, and
delete the crate. Also drop tonk-layout's dead tonk-concept dependency and
the stale Trunk watch path.
