--- rust/tonk-concept/Cargo.toml
@@ -1,54 +1,0 @@
-[package]
-name = "tonk-concept"
-version.workspace = true
-authors.workspace = true
-license.workspace = true
-edition.workspace = true
-description = "Shared notation-template machinery (source parsing, chrome/repeat plan, field substitution) backing <tonk-display>."
-
-[lib]
-crate-type = ["rlib"]
-
-[dependencies]
-indexmap = { workspace = true }
-ipld-core = { workspace = true }
-js-sys = { workspace = true }
-serde = { workspace = true, features = ["derive"] }
-serde-wasm-bindgen = { workspace = true }
-serde_json = { workspace = true }
-serde_ipld_dagjson = { workspace = true }
-tonk-schema = { workspace = true }
-wasm-bindgen = { workspace = true }
-web-sys = { workspace = true, features = [
-    "Attr",
-    "Comment",
-    "CustomElementRegistry",
-    "CustomEvent",
-    "CustomEventInit",
-    "Document",
-    "DocumentFragment",
-    "Element",
-    "HtmlElement",
-    "HtmlInputElement",
-    "HtmlScriptElement",
-    "HtmlTemplateElement",
-    "NamedNodeMap",
-    "Node",
-    "NodeFilter",
-    "NodeList",
-    "Text",
-    "TreeWalker",
-    "Window",
-] }
-
-[dev-dependencies]
-dialog-common = { workspace = true, features = ["helpers"] }
-tokio = { workspace = true, features = ["macros", "rt"] }
-wasm-bindgen-test = { workspace = true }
-
-[lints.rust]
-unexpected_cfgs = { level = "warn", check-cfg = [
-    'cfg(dialog_test_native_integration)',
-    'cfg(dialog_test_wasm_integration)',
-    'cfg(feature, values("web-integration-tests"))',
-] }
