+++ rust/tonk-ui/Trunk.toml
@@ -35,7 +35,6 @@ watch = [
     "../tonk-notation",
     "../tonk-board",
     "../tonk-workspace",
-    "../tonk-concept",
     "../tonk-display",
     "../tonk-host",
     "../tonk-layout",
